# go/model_location_details.go
/*
 * Ndscf_DataRepository
 *
 * Ndscf Data Repository Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved. 
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

type LocationDetails struct {

	AgeOfLocationInfo string `json:"ageOfLocationInfo,omitempty"`

	TrackingAreaId string `json:"trackingAreaId,omitempty"`

	PlmnId string `json:"plmnId,omitempty"`

	Lat string `json:"lat,omitempty"`

	Long string `json:"long,omitempty"`

	Elev string `json:"elev,omitempty"`
}

// AssertLocationDetailsRequired checks if the required fields are not zero-ed
func AssertLocationDetailsRequired(obj LocationDetails) error {
	return nil
}

// AssertRecurseLocationDetailsRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of LocationDetails (e.g. [][]LocationDetails), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseLocationDetailsRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aLocationDetails, ok := obj.(LocationDetails)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertLocationDetailsRequired(aLocationDetails)
	})
}