# go/model_sub_ids.go
/*
 * Ndscf_DataRepository
 *
 * Ndscf Data Repository Service. © 2021, 3GPP Organizational Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC). All rights reserved. 
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package datarepository

type SubIds struct {
	Items []string
}

// AssertSubIdsRequired checks if the required fields are not zero-ed
func AssertSubIdsRequired(obj SubIds) error {
	return nil
}

// AssertRecurseSubIdsRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of SubIds (e.g. [][]SubIds), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseSubIdsRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aSubIds, ok := obj.(SubIds)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertSubIdsRequired(aSubIds)
	})
}